Added PEP 8 Conventions with minor edits
We need to refactor code to fit PEP 8 code name convention

+++ Firmware/MSP430fr5994_MCU/msp430_proj.c
@@ -10,7 +10,7 @@
 
 /// Due to msp430.h, renamed msp430.c to msp430_proj.c to avoid any relation between the this file and msp430.h.
 
-int main(){
+short main(){
    mc_setup();
    struct MCScheduler mc_scheduler;
    buildScheduler(&mc_scheduler);
